+++ src/entities/problem/api/server/getDetailProblem.type.ts
