+++ src/entities/problem/index.ts
@@ -1,0 +1,4 @@
+export { getDetailProblem } from './api/server/getDetailProblem';
+export { getSessionKey } from './api/server/getSessionKey';
+
+export { default as DetailProblem } from './ui/DetailProblem';
