--- src/entities/problemSubmit/index.ts
@@ -1,1 +1,0 @@
-//
