--- src/entities/problemSubmit/model/mutations/submitCode.mutation.type.ts
@@ -1,5 +1,0 @@
-export interface ISubmitCodeRequest {
-  languageId: number;
-  sourceCode: string;
-  sessionKey: string;
-}
