--- src/features/problem/index.ts
@@ -1,7 +1,0 @@
-/*ui*/
-/**@description 코드 에디터, 터미널, 터미널패널이 있는 ProblemWorksSection 컴포넌트 */
-export { default as ProblemWorksSection } from './ui/ProblemWorksSection';
-export { default as DetailProblem } from './ui/DetailProblem';
-
-//actions
-export { getDetailProblem } from './api/server/getProblem';
