+++ src/features/submitCode/gitPush/lib/useGitPush.ts
@@ -1,8 +1,8 @@
 import {
   useGitPushAutoToggleMutation,
   useGitRepoChoice,
-} from '@/entities/submitProblem/gitpush/model/mutation/gitpush.mutation';
-import { useGetGitHubRepo } from '@/entities/submitProblem/gitpush/model/query/gitpush.query';
+} from '@/entities/submitCode/gitpush/model/mutation/gitpush.mutation';
+import { useGetGitHubRepo } from '@/entities/submitCode/gitpush/model/query/gitpush.query';
 import { useEffect, useState } from 'react';
 
 export default function useGitPush() {
