--- src/features/submitProblem/index.ts
@@ -1,5 +1,0 @@
-//actions
-export { getGitHubUrl } from './gitPush/api/server/getGitHub';
-
-//ui
-export { default as GitPushDialog } from './gitPush/ui/GitPushDialog';
