+++ src/shared/types/pathOptions.ts
@@ -1,1 +1,1 @@
-export type TGetProblemIdOptions = 'discussions' | 'submit-ready' | 'submit-prepare';
+export type TGetProblemIdOptions = 'discussions' | 'submit-ready' | 'submit-prepare' | 'review';
